Split multiprocessing tests
